Made reorganizational changes to support PyPI publishing of package
- Package prefix changed from `conjur_api_python3` to `conjur`
- Bundle name changed from `conjur-api-python` to `conjur-api`
- CLI renamed from `conjur-py3-cli` to `conjur-cli`
- Updated README to refer to new info
- New version bump done
