+++ pkg_bin/conjur-cli
@@ -13,7 +13,7 @@ sys.path.append(".")
 sys.path.append("..")
 
 #pylint: disable=wrong-import-position
-from conjur_api_python3.cli import Cli
+from conjur.cli import Cli
 
 if __name__ == '__main__':
     Cli.launch()
