+++ test/test_version.py
@@ -1,6 +1,6 @@
 import unittest
 
-from conjur_api_python3.version import __version__
+from conjur.version import __version__
 
 class VersionTest(unittest.TestCase):
     def test_version_is_set(self):
